more things split from steam.py to make reading the emulator a little less frustrating
